examples/use-case-3-keycloak-auth/implementation/services/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
httpx==0.26.0
pydantic==2.5.3
